requirements.txt
pyyaml
shell.py
pexpect
requests
pycrypto
